+++ src/ui/app_entry_template.js
