+++ src/ui/ui_bundle_collection.js
@@ -8,8 +8,8 @@ let unlink = promisify(require('fs').unlink);
 let readdir = promisify(require('fs').readdir);
 let readSync = require('fs').readFileSync;
 
-let UiBundle = require('./UiBundle');
-let appEntryTemplate = require('./appEntryTemplate');
+let UiBundle = require('./ui_bundle');
+let appEntryTemplate = require('./app_entry_template');
 
 class UiBundleCollection {
   constructor(bundlerEnv, filter) {
